test(crypto): split homoglyph discovery test into its own suite
The end-to-end homoglyph test pushed CryptoTokenDiscoveryServiceTests over
the SwiftLint type_body_length limit (264 > 250). Move it into a dedicated
@Suite file (CryptoTokenDiscoveryHomoglyphTests) rather than shrinking the
existing body. Behaviour unchanged; format-check green.
